chore: restructure styles
