+++ src/components/nav-item-group.css.ts
@@ -1,5 +1,5 @@
 import { style, styleVariants } from "@vanilla-extract/css"
-import { theme } from "../theme.css"
+import { theme } from "../styles/theme.css"
 import { media } from "./ui.css"
 import { mobileNavLink } from "./header.css"
