+++ src/templates/blog-post.css.ts
@@ -1,5 +1,5 @@
 import { style, globalStyle } from "@vanilla-extract/css"
-import { theme } from "../../src/theme.css"
+import { theme } from "../../src/styles/theme.css"
 
 export const blogPost = style({
   fontSize: theme.fontSizes[3],
